add compat test
